Renamed schedule app to subject
