+++ app/app/settings.py
@@ -50,7 +50,7 @@ INSTALLED_APPS = [
     "rest_framework.authtoken",
     "drf_spectacular",
     "core",
-    "schedule",
+    "selection",
     "subject",
     "user",
 ]
