+++ app/app/urls.py
@@ -22,14 +22,16 @@ from drf_spectacular.views import (
 )
 
 urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('api/schema/', SpectacularAPIView.as_view(), name='api-schema'),
+    path("admin/", admin.site.urls),
+    path("api/schema/", SpectacularAPIView.as_view(), name="api-schema"),
     path(
-        'api/docs/',
-        SpectacularSwaggerView.as_view(url_name='api-schema'),
-        name='api-docs'
+        "api/docs/",
+        SpectacularSwaggerView.as_view(url_name="api-schema"),
+        name="api-docs",
     ),
     path("api/users/", include("user.urls"), name="users-resource"),
     path("api/subjects/", include("subject.urls"), name="subject-resource"),
-    path("api/schedules/", include("schedule.urls"), name="schedule-resource"),
+    path(
+        "api/selections/", include("selection.urls"), name="selection-resource"
+    ),
 ]
