+++ app/selection/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class ScheduleConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "schedule"
+    name = "selection"
