+++ app/selection/urls.py
@@ -1,13 +1,16 @@
-from schedule import views
+from selection import views
 from rest_framework.urls import path
 
-app_name = "schedule"
+app_name = "selection"
 
 
 urlpatterns = [
-    path('', views.SelectionListView.as_view(), name='selection-list'),
-    path('<str:id>', views.SelectionDetailView.as_view(),
-         name='selection-detail'),
+    path("", views.SelectionListView.as_view(), name="selection-list"),
+    path(
+        "<str:id>",
+        views.SelectionDetailView.as_view(),
+        name="selection-detail",
+    ),
     path(
         "<uuid:selection_id>/subjects/",
         views.SubjectSectionListView.as_view(),
@@ -17,5 +20,5 @@ urlpatterns = [
         "<uuid:selection_id>/subjects/<int:subject_section_id>/",
         views.SubjectSectionDetailsView.as_view(),
         name="subject-details",
-    )
+    ),
 ]
